requirement.txt
fastapi
uvicord
dnspython